requirements.txt
asgiref==3.2.3
certifi==2019.6.16
chardet==3.0.4
Django==3.0.14
djangorestframework==3.11.2
idna==2.8
Pillow==8.1.1
pytz==2019.1
sqlparse==0.3.0
django.filter==2.4.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1 
